--- frontend/src/app/login/components/signup/signup.component.html
@@ -1,21 +1,0 @@
-<div class="up">
-  <form id="signup" name="signup" method="post" action="/signup">
-      <label for="email">Email Address
-          <input class="text" name="email" type="email">
-      </label>
-
-      <label for="firstname">Firstname
-          <input name="firstname" type="text">
-      </label>
-
-      <label for="lastname">Lastname
-          <input name="lastname" type="text">
-      </label>
-
-      <label for="password">Password
-          <input name="password" type="password">
-      </label>
-
-      <input class="btn" type="submit" value="Sign Up">
-  </form>
-</div>
